Factoring service integration tests
