+++ service/src/test/java/org/rocklass/fullstacklab/service/ItemControllerIT.java
@@ -1,41 +1,18 @@
 package org.rocklass.fullstacklab.service;
 
-import java.net.URL;
-
-import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.rocklass.fullstacklab.Application;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.test.IntegrationTest;
-import org.springframework.boot.test.SpringApplicationConfiguration;
-import org.springframework.boot.test.TestRestTemplate;
+import org.rocklass.fullstacklab.test.ServiceIntegrationTest;
 import org.springframework.http.ResponseEntity;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.springframework.test.context.web.WebAppConfiguration;
-import org.springframework.web.client.RestTemplate;
 
-@RunWith(SpringJUnit4ClassRunner.class)
-@SpringApplicationConfiguration(classes = Application.class)
-@WebAppConfiguration
-@IntegrationTest({"server.port=0"})
-public class ItemControllerIT {
-
-    @Value("${local.server.port}")
-    private int port;
-
-	private URL base;
-	private RestTemplate template;
-
-	@Before
-	public void setUp() throws Exception {
-		this.base = new URL("http://localhost:" + port + "/items");
-		template = new TestRestTemplate();
+public class ItemControllerIT extends ServiceIntegrationTest {
+	@Override
+	public String getRequestMapping() {
+		return "/items";
 	}
 
 	@Test
 	public void getItems() throws Exception {
-		ResponseEntity<String> response = template.getForEntity(base.toString(), String.class);
+		ResponseEntity<String> response = getTemplate().getForEntity(getBase().toString(), String.class);
 		// TODO write IT
 	}
 }
